Sentries module is now error free

+++ src/sentry/icbm/sentry/turret/sentries/TileEntityGunTurret.java
@@ -32,7 +32,7 @@ public class TileEntityGunTurret extends TileEntityAutoTurret
     }
 
     @Override
-    public float getFiringRequest()
+    public long getFiringRequest()
     {
         return 1;
     }
